assignment4/requirements.txt
"""
===============================================================================
FILE 3: requirements.txt
===============================================================================
"""

# requirements.txt
kafka-python==2.0.2
matplotlib==3.7.1
numpy==1.24.3

"""